+++ tests/test_checker/test_noqa.py
@@ -122,6 +122,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS346': 1,
     'WPS347': 3,
     'WPS348': 1,
+    'WPS349': 1,
 
     'WPS400': 0,
     'WPS401': 0,
@@ -172,6 +173,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS446': 1,
     'WPS447': 1,
     'WPS448': 1,
+    'WPS449': 1,
 
     'WPS500': 1,
     'WPS501': 1,