+++ CHANGELOG.md
@@ -17,14 +17,23 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids using nested `typing.Literal`, `typing.Union` and `typing.Annotated`
 - Forbids use of vague import names (e.g. `from json import loads`)
 - Make ``OveruseOfNoqaCommentViolation`` configurable (`--max-noqa-comments`)
+- Improves tests: we now ensure that each violation with previous codes also 
+  has corresponding versions changed in their documentation
 - Forbid incorrectly swapped variables
 - Forbids to use `+=` with list arguments
+- Forbids to use redundant subscripts (e.g., `[0:7]` or `[3:None]`)
 
 ### Bugfixes
 
 - Fixes ``ImplicitElifViolation`` false positives on a specific edge cases
 - Fixes ``I_CONTROL_CODE setting`` for ``BadMagicModuleFunctionViolation``
+- Fixes compatibility with flake8 `3.8.x`
+- Fixes that `not not True` was not detected as `WPS330`
+- Fixes addition of ``MisrefactoredAssignmentViolation`` check
 
+### Misc
+
+- Replace `scripts/tokens.py` with an external tool.
 
 ## 0.12.5